+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 This release has a basic approval process for a single organization on board.
 
-- `ADD` Message details can be viewed
+- `ADD` Message details can be viewed for approvals
+- `ADD` Message details can be viewed for my organization
 - `ADD` Rejected message can have a reason
 - `ADD` Messages waiting for approval can be rejected
 - `ADD` Dev script has a task to reset the task database